fix LibReadProvider.kt
